pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hnh"
version = "0.1.0"
description = "HnH — deterministic personality simulation engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "structlog>=24.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
astrology = [
    "pyswisseph>=2.10",
]

[project.scripts]
hnh = "hnh.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hnh*"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = "/(\\.git|\\.venv|__pycache__)/"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --strict-markers"
markers = [
    "unit: unit tests",
    "integration: integration tests",
]

[tool.coverage.run]
source = ["hnh"]
branch = true
omit = ["hnh/interface/llm_adapter.py", "tests/*"]

[tool.coverage.report]
fail_under = 90
show_missing = true
